+++ changelog.md
@@ -1,3 +1,9 @@
+# v2.2.1
+
+## Changes
+
+Updated for patch 10.0.7.
+
 # v2.2.0
 
 ## Changes